requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
faiss-cpu>=1.7.4
rank_bm25>=0.2.2
tavily-python>=0.2.9
fastapi>=0.109.0
uvicorn>=0.25.0
pydantic>=2.5.0
tiktoken>=0.5.2
httpx>=0.26.0
python-dotenv>=1.0.0
openai>=1.0.0